+++ src/parser/inline.ts
@@ -12,6 +12,7 @@ import { html } from './inline/html';
 import { insertion } from './inline/insertion';
 import { deletion } from './inline/deletion';
 import { mark } from './inline/mark';
+import { emstrong } from './inline/emstrong';
 import { emphasis } from './inline/emphasis';
 import { strong } from './inline/strong';
 import { code } from './inline/code';
@@ -35,6 +36,7 @@ export import HTMLParser = InlineParser.HTMLParser;
 export import InsertionParser = InlineParser.InsertionParser;
 export import DeletionParser = InlineParser.DeletionParser;
 export import MarkParser = InlineParser.MarkParser;
+export import EmStrongParser = InlineParser.EmStrongParser;
 export import EmphasisParser = InlineParser.EmphasisParser;
 export import StrongParser = InlineParser.StrongParser;
 export import CodeParser = InlineParser.CodeParser;
@@ -60,6 +62,7 @@ export const inline: InlineParser = lazy(() => union([
   insertion,
   deletion,
   mark,
+  emstrong,
   strong,
   emphasis,
   code,