+++ src/parser/inline/autolink/hashnum.test.ts
@@ -10,6 +10,7 @@ describe('Unit: parser/inline/autolink/hashnum', () => {
       assert.deepStrictEqual(inspect(parser('')), undefined);
       assert.deepStrictEqual(inspect(parser('#')), [['#'], '']);
       assert.deepStrictEqual(inspect(parser('# ')), [['#'], ' ']);
+      assert.deepStrictEqual(inspect(parser('#1_')), [['#1_'], '']);
       assert.deepStrictEqual(inspect(parser('#1#')), [['#1#'], '']);
       assert.deepStrictEqual(inspect(parser('#1#2')), [['#1#2'], '']);
       assert.deepStrictEqual(inspect(parser('#1#2#3')), [['#1#2#3'], '']);
@@ -32,19 +33,18 @@ describe('Unit: parser/inline/autolink/hashnum', () => {
       assert.deepStrictEqual(inspect(parser('0##a')), [['0##a'], '']);
       assert.deepStrictEqual(inspect(parser('あ#1')), [['あ#1'], '']);
       assert.deepStrictEqual(inspect(parser(' #1')), undefined);
-      assert.deepStrictEqual(inspect(parser('#12345678901234567')), [['#12345678901234567'], '']);
-      assert.deepStrictEqual(inspect(parser(`#${'1'.repeat(16)}a`)), [[`#${'1'.repeat(16)}a`], '']);
+      assert.deepStrictEqual(inspect(parser(`#${'1'.repeat(10)}`)), [[`#${'1'.repeat(10)}`], '']);
+      assert.deepStrictEqual(inspect(parser(`#${'1'.repeat(10)}a`)), [[`#${'1'.repeat(10)}a`], '']);
     });
 
     it('valid', () => {
       assert.deepStrictEqual(inspect(parser('#1')), [['<a class="hashnum">#1</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('#1_')), [['<a class="hashnum">#1</a>'], '_']);
       assert.deepStrictEqual(inspect(parser('#1 ')), [['<a class="hashnum">#1</a>'], ' ']);
       assert.deepStrictEqual(inspect(parser('#1\n')), [['<a class="hashnum">#1</a>'], '\n']);
       assert.deepStrictEqual(inspect(parser('#1\\')), [['<a class="hashnum">#1</a>'], '\\']);
       assert.deepStrictEqual(inspect(parser('#1\\ ')), [['<a class="hashnum">#1</a>'], '\\ ']);
       assert.deepStrictEqual(inspect(parser('#1\\\n')), [['<a class="hashnum">#1</a>'], '\\\n']);
-      assert.deepStrictEqual(inspect(parser('#1234567890123456')), [['<a class="hashnum">#1234567890123456</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('#123456789')), [['<a class="hashnum">#123456789</a>'], '']);
     });
 
   });