+++ src/parser/inline/emphasis.ts
@@ -1,6 +1,8 @@
 import { EmphasisParser } from '../inline';
 import { union, some, syntax, surround, open, lazy } from '../../combinator';
 import { inline } from '../inline';
+import { emstrong } from './emstrong';
+import { strong } from './strong';
 import { str } from '../source';
 import { Syntax, State } from '../context';
 import { startTight, blankWith } from '../visibility';
@@ -8,12 +10,17 @@ import { unshift } from 'spica/array';
 import { html, defrag } from 'typed-dom/dom';
 
 export const emphasis: EmphasisParser = lazy(() => surround(
-  str('_', '_'),
+  str('*', '*'),
   syntax(Syntax.none, 1, 1, State.none,
   startTight(some(union([
-    some(inline, blankWith('_'), [[/^\\?\n/, 9]]),
-    open(some(inline, '_', [[/^\\?\n/, 9]]), emphasis),
+    strong,
+    some(inline, blankWith('*'), [[/^\\?\n/, 9]]),
+    open(some(inline, '*', [[/^\\?\n/, 9]]), union([
+      emstrong,
+      strong,
+      emphasis,
+    ])),
   ])))),
-  str('_'), false,
+  str('*'), false,
   ([, bs], rest) => [[html('em', defrag(bs))], rest],
   ([as, bs], rest) => [unshift(as, bs), rest]));