+++ src/parser/inline/strong.ts
@@ -1,6 +1,7 @@
 import { StrongParser } from '../inline';
 import { union, some, syntax, surround, open, lazy } from '../../combinator';
 import { inline } from '../inline';
+import { emstrong } from './emstrong';
 import { str } from '../source';
 import { Syntax, State } from '../context';
 import { startTight, blankWith } from '../visibility';
@@ -8,12 +9,15 @@ import { unshift } from 'spica/array';
 import { html, defrag } from 'typed-dom/dom';
 
 export const strong: StrongParser = lazy(() => surround(
-  str('*', '*'),
+  str('**', '*'),
   syntax(Syntax.none, 1, 1, State.none,
   startTight(some(union([
-    some(inline, blankWith('*'), [[/^\\?\n/, 9]]),
-    open(some(inline, '*', [[/^\\?\n/, 9]]), strong),
+    some(inline, blankWith('**'), [[/^\\?\n/, 9]]),
+    open(some(inline, '*', [[/^\\?\n/, 9]]), union([
+      emstrong,
+      strong,
+    ])),
   ])))),
-  str('*'), false,
+  str('**'), false,
   ([, bs], rest) => [[html('strong', defrag(bs))], rest],
   ([as, bs], rest) => [unshift(as, bs), rest]));